Fix migration

+++ gouden-draak/database/migrations/2021_06_05_082542_create_gerecht_table.php
@@ -19,9 +19,9 @@ class CreateGerechtTable extends Migration
             $table->tinyText('menu_toevoeging')->nullable();
             $table->mediumText('naam');
             $table->decimal('prijs', 10, 2);
-            $table->integer('soortgerecht_id');
+            $table->unsignedBigInteger('soortgerecht_id');
             $table->longText('beschrijving');
-            $table->integer('pittigheid_id');
+            $table->unsignedBigInteger('pittigheid_id');
             $table->foreign('soortgerecht_id')->references('id')->on('soort_gerecht');
             $table->foreign('pittigheid_id')->references('id')->on('pittigheids');
         });
